requirements.txt
# =========================================================
# Sistema RAG Avanzado - Dependencias
# =========================================================

# --- API Framework ---
fastapi==0.115.6
uvicorn[standard]==0.32.1

# --- Validación de datos ---
pydantic==2.10.3
pydantic-settings==2.7.0

# --- Variables de entorno ---
python-dotenv==1.0.1

# --- LangChain Core ---
langchain==0.3.13
langchain-core==0.3.28
langchain-community==0.3.13

# --- LLM: Google Gemini (Google AI Studio) ---
langchain-google-genai==2.0.7
google-generativeai==0.8.3

# --- LLM: Azure OpenAI (alternativa) ---
langchain-openai==0.2.14
openai==1.58.1

# --- Embeddings y Vector Store ---
faiss-cpu==1.9.0
langchain-huggingface==0.1.2
sentence-transformers==3.3.1

# --- Procesamiento de documentos ---
pypdf==5.1.0
python-docx==1.1.2
python-multipart==0.0.20

# --- Utilidades ---
aiofiles==24.1.0
